+++ src/graphql/resolver/create_resolver_function/get_resolver_input/mod.rs
@@ -3,7 +3,10 @@ use bson::Document;
 use log::debug;
 
 use crate::{
-    configuration::subgraph::entities::service_entity_field::ServiceEntityFieldConfig,
+    configuration::subgraph::entities::{
+        service_entity_field::ServiceEntityFieldConfig, ServiceEntityConfig,
+    },
+    data_sources::DataSources,
     graphql::schema::ResolverType,
 };
 
@@ -19,8 +22,11 @@ impl ServiceResolver {
         ctx: &ResolverContext,
         as_type_field: &Option<ServiceEntityFieldConfig>,
         resolver_type: &ResolverType,
+        data_sources: &DataSources,
+        entity: &ServiceEntityConfig,
     ) -> Result<Option<Document>, async_graphql::Error> {
         debug!("Getting Resolver Input");
+        let data_source = DataSources::get_entity_data_soruce(data_sources, entity);
         let input_document = match resolver_type {
             ResolverType::InternalType => {
                 let as_field = if let Some(as_field) = as_type_field {
@@ -28,7 +34,7 @@ impl ServiceResolver {
                 } else {
                     return Err(async_graphql::Error::new("Field is not a nested field"));
                 };
-                ServiceResolver::create_internal_input(&ctx, as_field.clone())?
+                ServiceResolver::create_internal_input(&ctx, as_field.clone(), data_source)?
             }
             _ => {
                 let input = ctx.args.try_get(&format!("{}_input", ctx.field().name()))?;